requirements.txt
attrs==19.3.0
bcrypt==3.1.7
cffi==1.13.2
cryptography==2.8
dnspython==1.16.0
falcon==2.0.0
gunicorn==20.0.4
importlib-metadata==1.3.0
jsonschema==3.2.0
marshmallow==3.3.0
more-itertools==8.0.2
munch==2.5.0
phonenumbers==8.11.1
pycountry==19.8.18
pycparser==2.19
PyJWT==1.7.1
pymodm==0.4.2
pymongo==3.10.0
pyrsistent==0.15.6
python-dateutil==2.8.1
python-dotenv==0.10.3
six==1.13.0
validator-collection==1.4.0
zipp==0.6.0
